Add additional utils tests + big re-organisation of utils file structure
