--- src/utils/index.ts
@@ -1,2 +1,0 @@
-export { extract } from './extract'
-export { assign, type AssignInput, type AssignOptions } from './assign'
